Using progress spinners & clean output/status system.
Any active builds, and asset pipelines will trigger a progress spinner
on the CLI to show that work is in-progress.

All other output us using this system to ensure that output works
together properly with the progress system.

Using `console::Emoji` to conditionally print emojis when they are
supported in the runtime environment.

closes #1