Major refactor of code layout and structure.
This should allow for a much easier process of adding new pipelines and
assets to the mix. Functionality has been preserved. A few small bug
fixes have been factored in.

Add a Trunk, Yew, YBC (Bulma) example. This includes a minimal
Trunk.toml.

Add snapshot tests over the yew example app.

Added a pull request template, just to help folks remember to knock out
a few of the small tasks associated with getting PRs to land.

Broke apart the config.rs module into a directory of a few separate
modules under src/config. The file was getting fairly large and had
three distinct families of types.

Simplified the handling of errors in main. This handles returning proper
status codes and also prints out the chain of errors associated with the
error being logged.

The processing of Trunk.toml files has been updated to ensure that all
paths declared in the file will be treated as relative to the config
file itself (if the paths are indeed relative).

Introduced CI caching.

closes #58