Major refactor of pipeline system to support next wave of features.
All assets which are to be processed by trunk must now be declared as
HTML `link` elements as such:
`<link data-trunk rel="rust|sass|..." data-attr0 data-attr1/>`.
The links may appear anywhere in the HTML and Trunk will process them and
replace them or delete them based on the associated pipeline's output.
If the link element does not have the `data-trunk` attribute, it will not
be processed.

The `--manifest-path` flag has been removed in favor of the new link
style: `<link rel="rust" href="path/to/Cargo.toml"/>`. This is optional,
and will default to the Cargo.toml of the source HTML's parent dir if
not specified.

Update README describing new asset pipeline declaration pattern.

closes #50