Added the inline pipeline & updated CHANGELOG.md and README.md
The inline pipeline allows including the content of files directly into `index.html`. Just add `<link data-trunk rel="inline" href="path/to/file">` to the position in your `index.html` where the content of the files should be placed, and you're good to go. There three content types available: `html`, `css`, and `js`. `html` is pasted into `index.html` as is, `css` is wrapped in `style` tags and `js` in wrapped in `script` tags.
The type can also be specified with the `type` attribute.