A few small tweaks to wasm-opt declaration for consistency.
This updates the rel="rust" asset link to expect `data-wasm-opt` instead
of just `wasm-opt`. I do actually prefer the latter, having to prefix
with `data-` is a bit of an annoyance. However, it is what we are
currently doing everywhere else, and I would prefer to keep it
consistent for now.

Perhaps we can get some feedback from the community on whether or not
they perceive these `data-` prefixes as important. The HTML is processed
and those links are typically entirely removed; however, it is
"semantic" HTML to prefix such attrs with `data-`.