Made the first step towards a plugin system for trunk
The intend of this commit is not to deliver anything near a finished solution. The idea is to provide a building block and a place for experimenting with different ideas.

What changed:
  - trunk is now a workspace
  - there are now two additional crates: trunk-derive and trunk-plugin

The purpose of trunk-plugin is to become the library used by plugins and trunk itself. It provides two main data structures: Args and Output. Args it the struct for passing arguments to an extension, while Output is the struct used by plugins to return something.

The purpose of trunk-derive is to make the life for plugin authors easier, by removing some boilerplate.

Both libraries are also responsible for hiding away the complexity that currently comes from passing complex data structures between Rust and WASM.
When a WASM function wants to receive a struct from Rust, the Rust code first has to serialize the struct to bytes, then store these bytes in the WASM module's memory, and finally pass both a pointer and the length of the memory slice to WASM. The WASM function then has to dereference the pointer (unsafe) and deserialize the bytes again.
When the WASM function then wants to return a struct to Rust, it also has to serialize the struct to bytes and send both a pointer and the length of the memory slice to Rust. The Rust code then has to read the bytes from the WASM module's memory and deserialize them.

This is a lot of boilerplate and a lot of potential pitfalls, so plugin authors should have to handle that themself.