Added the skeleton for the plugin pipeline
Invoking plugins can be done the same way as invoking any other kind of pipeline. This has the advantage that it's quite easy to integrate.

One question that remains however is, if plugins should inherit the pipeline design by exposing a `run` and a `finalize` function, or if plugins should have only one `main` function. The advantage of a single main function is simplicity on the plugin side. The disadvantage is that this makes integrating the plugin pipeline harder, since the plugin might need to read the index html document, which usually isn't handed to the pipeline before `finalize`.