Update & isolate version checking & formatting of version output
This changeset isolates the code which defines the version test to use
for various downloadable applications, and also isolates the code used
to format the version output. This allows for easier unit testing of
this logic.

Added unit tests for variations of wasm-bindgen & wasm-opt version
output which differs based on installation of pre-compiled binaries vs
building from source.