Shut down proxied websockets more promptly
Before, it would wait for both the `stream_to_frontend` and
`stream_to_backend` futures to complete before shutting down the socket,
but once either side is broken there's no point in continuing with
the other.

In particular, this allows the browser to immediately notice when a
backend server is restarted, instead of delaying until it tries to
send a message.