Prevent buildup of change events, debouncing while busy
Changes from the file system already get debounced in a 1 second
window. However, still each change that gets reported, will trigger a
build. All changes happening during the build will get buffered and
sequentially be executed.

Depending on the length of the build, and the input from the user
during this time, this can lead to a huge amount of changes, leading
to a buildup that makes it look like trunk is "continuously"
rebuilding.

This change debounces change event while the build is active. It will
record the request for a new build, and notify the build loop. Only
when the build loop has started can another request be recorded.

This results in a build starting as soon as possible when the first
events emitted and ensures at most one more run after the build
has finishes. No additional requests will be queued.