feat: bring back the fix for queuing up of change events
While the change was merged in the past, it got removed a few commits
later. This leads to the problem that disabling cooldown brings back
the problem of piling up events, while enabling cooldown tends to
not update/trigger the build at all.

This change brings back the logic which sums up all events happening
*during* a build to a single build trigger after the build and disables
the cooldown by default (it's opt-in now).