watch: match on `All` event
Some platforms don't support precise notification kinds and instead
report `All` as an event kind. This is what the notification system
reports when it doesn't know what else to report.

It's also the kind that's used in "imprecise" mode.

This fixes watching on Windows, and may address issues reported in #232.

Signed-off-by: Sean Cross <sean@xobs.io>