fix: Treat errors as errors when detecting a cargo project
So far, when a cargo project was implicitly configured failing to parse
the Cargo.toml or loading the metadata resulted in the state of
"no rust project found". This is actually wrong, as a project was there,
found, but it metadata could not be evaluated.

This can lead to silent build failures, resulting in broken build
artifacts.

This change only considers the fact that a Cargo.toml is missing as an
indication that this is a non-Rust project. If a Cargo.toml file is
present (or it's presence cannot be evaluated) and it fails to generate
the metadata, it will be considered a build failure.

closes #11