fix: handle projects which have both a bin and lib (cdylib) target
In a previous change `cdylib` targets have been added to the list of
allowed artifacts. However, projects which are both a binary as well as
a library, the compilation failed now, since both target artifacts (bin
and lib) got added to the target output. It was not possible to select
by name.

This change adds the `data-target-name` attribute to specifically select
a target "by name", without having to use the `--bin` argument in cargo.

Closes: #630