fix: await all asset pipeline, even if some fail
When a pipeline step fails, the while loop will fail fast, but this
will leave all other tasks still running in the background. Which might
lead to the case where longer running tasks (like cargo build) run,
while others abort the build.

Having a case like this, might also lead to multiple cargo builds
stacking up in the background.

It also might lead to the case where a failure, due to a missing
directory, will repeated, as repeated builds are triggered by the
initial cargo build. Because only after the initial cargo build, the
target folder will be ignored. But as the build fails fast, a new build
will be triggered, as changes to the target folder are not yet ignored.

Closes: #662