chore: clean up a few things around the script element check
Rename the flag to `allow_self_closing_script`, which may be verbose,
but it's also not recommended to use anyway. And it feels a bit more
descriptive, we don't ignore script errors, but allow the error of a
self-closing script tag.

Ensure that the configuration from the config file as well as from the
CLI gets evaluated.

Introduce an "options" struct, making future enhancements a bit simpler.

Use plain strings for errors, vs concatenating and formatting.