feat: provide total wasm size to the initializer
The initializer tries to get the total WASM size from the response.
This can be done using the `Content-Length` header, but that approach
has some downsides. First of all, it's not required and my be missing.

Second, it will report the "on the wire" size of the response, the
stream which already content encoded. If that includes compression,
then there is no way to evaluate the total (uncompressed) size, or
the currently read (compressed) bytes. So there is no way to sum up
the bytes read (decoded, uncompressed) towards the total (compressed).

On the other side, during the bundling process, we do know the total
WASM size, and we also create hashes for it, so that we ensure its
integrity. So we can also use that size as a total.