fix: retain the insertion order of SRI resources
This fixes an oversight where the order of <link rel=preload> statements
was different than the order of insertion into the SriBuilder.

This was caused by the usage of BTreeMap, which internally reordered the
entries in lexicographic order - Preload types first, then
ModulePreloads, sorted by the lexicographic ordering of strings in case
of any conflicts. This was likely not the intended result.

This issue was mitigated by replacing the BTreeMap with a Vec, and
manually implementing checks for identical keys.