feat: Rework the configuration system
This refactors most parts of the configuration system. Splitting it
up into a command line argument model, a persistent config file model,
and the runtime model.

The command line arguments (including env-vars) will override the
persisted (or default) config model.

It also adds the ability to parse YAML, JSON, and cargo metadata
in addition to the TOML variant.

This prepares the configuration for future enhancements, like build
profiles, ...