add proxy option `no_redirect`
When serving responses with the Trunk proxy, redirects are automatically
followed by default. This can be undesirable, e.g. when serving a login page
that response with a redirect and a set-cookie header like

```
HTTP/1.1 303 See Other
location: /
set-cookie: session=...; ...;
```

This PR adds a proxy option `no_redirect` that when set will instruct the proxy
to not automatically follow redirects. The default behavior is unchanged.