feat: Initial support for cargo profiles
This isn't the complete "trunk profile" implementation, but maybe a
short term solution in the right direction.

This allows setting a cargo profile on multiple levels. The idea is to
allow influencing the cargo profile only. The trunk release flag still
works as before. However you can now choose a `--cargo-profile` from
the command line, from the Trunk.toml, from the `index.html` data
attributes. In that order of preference.