docs: shared webworker example
Added a simple SharedWorker example based on webworker-module example.
It demonstrates using both set_on<event> and addEventListener
interfaces.

As of this moment this example has a problem to be fixed where
the first page which creates a shared worker does not reach the
onconnect handler, so it will not receive messages from worker nor be
able to post messages to unless it creates a new SharedWorker object
which will this time reach already initialized
SharedWorkerGlobalScope's onconnect.